+++ src/ops/rms_norm.h
@@ -10,7 +10,7 @@ class RMSNorm : public Op
 {
 public:
   RMSNorm (GPUDevice *dev, Command *command, Tensor weight,
-           const float eps = 1e-3, const Tensor::DType dtype = Tensor::FP32);
+           const float eps = 1e-3, const Tensor::DType dtype = Tensor::FP16);
   absl::StatusOr<Tensor> operator() (Tensor a) noexcept;
   absl::Status init () noexcept override;
   uint64_t time () noexcept override;
